Send booking info by email
